Refactored unwieldy jinja.py
